# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Decision Routing System API
  version: 0.1.0
paths:
...
  /auth/me:
    get:
      tags:
      - Auth
      summary: Получить текущего пользователя
      description: Возвращает пользователя, найденного через Basic Auth/Active Directory, его роль, отдел,
        должность и права frontend.
      operationId: get_current_user_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
      security:
      - HTTPBasic: []
...
  /priority-settings:
    get:
      tags:
      - Priority
      summary: Получить коэффициенты расчета приоритета
      description: Обычный руководитель получает настройки только своего отдела в режиме чтения. top-manager
        получает все отделы и может редактировать. Дополнительно отдаёт read-only параметры срочности
        (urgent) для предпросмотра на фронте.
      operationId: get_priority_settings_priority_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrioritySettingsResponse'
      security:
      - HTTPBasic: []
    put:
      tags:
      - Priority
      summary: Сохранить коэффициенты расчета приоритета
      description: Доступно только top-manager.
      operationId: update_priority_settings_priority_settings_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrioritySettingsModel'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrioritySettingsModel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBasic: []
...
  /analytics/applications:
    get:
      tags:
      - Analytics
      summary: Аналитика по заявкам
      operationId: analytics_applications_analytics_applications_get
      security:
      - HTTPBasic: []
      parameters:
      - name: createdFrom
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdfrom
      - name: createdTo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdto
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics/executors:
    get:
      tags:
      - Analytics
      summary: Аналитика по исполнителям
      operationId: analytics_executors_analytics_executors_get
      security:
      - HTTPBasic: []
      parameters:
      - name: createdFrom
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdfrom
      - name: createdTo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdto
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics/work-types:
    get:
      tags:
      - Analytics
      summary: Аналитика по видам работ
      operationId: analytics_work_types_analytics_work_types_get
      security:
      - HTTPBasic: []
      parameters:
      - name: createdFrom
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdfrom
      - name: createdTo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdto
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics/departments:
    get:
      tags:
      - Analytics
      summary: Аналитика по отделам
      operationId: analytics_departments_analytics_departments_get
      security:
      - HTTPBasic: []
      parameters:
      - name: createdFrom
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdfrom
      - name: createdTo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdto
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    CurrentUserOut:
      properties:
        id:
          type: string
          title: Id
        login:
          type: string
          title: Login
        fullName:
          type: string
          title: Fullname
        roles:
          items:
            type: string
          type: array
          title: Roles
        departmentId:
          type: string
          title: Departmentid
        postName:
          type: string
          title: Postname
        positionId:
          type: string
          title: Positionid
        isActive:
          type: boolean
          title: Isactive
      type: object
      required:
      - id
      - login
      - fullName
      - roles
      - departmentId
      - postName
      - positionId
      - isActive
      title: CurrentUserOut
      description: Authenticated user (/auth/me). Carries the full set of system roles.
    CurrentUserResponse:
      properties:
        user:
          $ref: '#/components/schemas/CurrentUserOut'
        permissions:
          $ref: '#/components/schemas/UserPermissionsOut'
      type: object
      required:
      - user
      - permissions
      title: CurrentUserResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PrioritySettingsModel:
      properties:
        department:
          additionalProperties:
            type: number
          type: object
          title: Department
        managerAuthor:
          additionalProperties:
            type: number
          type: object
          title: Managerauthor
        deadline:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Deadline
      type: object
      required:
      - department
      - managerAuthor
      - deadline
      title: PrioritySettingsModel
    PrioritySettingsResponse:
      properties:
        department:
          additionalProperties:
            type: number
          type: object
          title: Department
        managerAuthor:
          additionalProperties:
            type: number
          type: object
          title: Managerauthor
        deadline:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Deadline
        urgent:
          $ref: '#/components/schemas/UrgentSettingsOut'
      type: object
      required:
      - department
      - managerAuthor
      - deadline
      - urgent
      title: PrioritySettingsResponse
...
    UrgentSettingsOut:
      properties:
        thresholdHours:
          type: number
          title: Thresholdhours
        bonus:
          type: number
          title: Bonus
      type: object
      required:
      - thresholdHours
      - bonus
      title: UrgentSettingsOut
      description: Параметры бонуса срочности (read-only, из config.json → priority).
...
    UserPermissionsOut:
      properties:
        canManageEmployees:
          type: boolean
          title: Canmanageemployees
        canManageWorkTypes:
          type: boolean
          title: Canmanageworktypes
        canManagePrioritySettings:
          type: boolean
          title: Canmanageprioritysettings
        canViewReports:
          type: boolean
          title: Canviewreports
      type: object
      required:
      - canManageEmployees
      - canManageWorkTypes
      - canManagePrioritySettings
      - canViewReports
      title: UserPermissionsOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic
...